src/hydraa_faas/agent/requirements.txt
fastapi
uvicorn
pydantic
PyYaml
httpx
kubernetes